wordcount: ignore case for words and use a t_scope.
Ignore the case for word equality in the map.
Use the t_scope for the vector instead of the libc heap.
Instead of keeping a reference to the word from the mmapped file
content, duplicate the words on the t_scope and convert them to lower
case in the vector.

* Ignore case for word equality for word occurrences map
* Create a t_scope in main()
* Use this t_scope for the vector
* Duplicate the words on the t_scope and convert them to lower case
* The words in the vector no longer reference the file content

Change-Id: Ic47d7a96625acb75237b3c897bb6fbe7c69c2efa