wordcount: split-up server and client part of wordcount.
We want the wordcount example to be multi-processes.
Split up the wordcount into server and client parts.
Add wordcount-base.h for common definitions.

Server:
Receive the file content from the client via RPC, count the number of
occurrences of each unique words, and send the results back.

Client:
Read the content of the file located at <file_path> and send it to the
server to get the number of occurrences of each unique words via RPC.

Change-Id: If95c6f7b4415c4e393b6e2ea5f76514a6ca3b08a