wordcount: add Python version of client wordcount-client.py.
The IOPs and RPCs can also be used in Python with IOPy.

IOPy is a Python module to be able to easily manipulate the IOPs and
call RPCs on other Intersec processes.
It is available in lib-common/src/iopy.

Add a Python version of wordcount client, wordcount-client.py, using
IOPy.

Change-Id: Id9ace11a1d6329ef28146a1784f3a20c26205fd9